DTComPair-dataset 1

Hypothetical data from a paired study that aims to compare the accuracy of two binary diagnostic tests.

Usage

Paired1

Format

A dataframe containing 3 columns (d, y1 and y2) and 712 rows.

d is a numeric vector specifying the gold-standard results (1 = presence of disease, 0 = absence of disease).

y1 is a numeric vector specifying the results of diagnostic test 1 (1 = positive, 0 = negative).

y2 is a numeric vector specifying the results of diagnostic test 1 (1 = positive, 0 = negative).

Examples

data(Paired1) # Hypothetical study data 
ftable(Paired1)
#>      y2   0   1
#> d y1           
#> 0 0     155  22
#>   1      53  31
#> 1 0      32  22
#>   1      78 319