Accuracy of Two Binary Diagnostic Tests in a Paired Study Design

Sensitivity and specificity, (positive and negative) predictive values and (positive and negative) diagnostic likelihood ratios of a two binary diagnostic tests in a paired study design.

Usage

acc.paired(tab, alpha, method.ci, ...)

Arguments

tab

An object of class tab.paired.

alpha

Significance level alpha for 100(1-alpha)%-confidence intervals, the default is 0.05.

method.ci

A function used to compute the confidence intervals for sensitivity, specificity, and predictive values. The default is waldci for Wald's asymptotic normal-based confidence intervals. See acc.1test.

...

Additional arguments, usually not required.

Details

The calculation of accuracy measures follows standard methodology, e.g. described in Pepe (2003) or Zhou et al. (2011).

The confidence intervals for sensitivity, specificity, and predictive values are computed using the methodology implemented in the function passed to the argument method.ci.

Confidence intervals for diagnostic likelihood ratios are computed according to Simel et al. (1991).

Value

A list of class acc.paired:

Test1

A list of class acc.1test containing results and accuracy estimates of Test 1.

Test2

A list of class acc.1test containing results and accuracy estimates of Test 2.

References

Pepe, M. (2003). The statistical evaluation of medical tests for classifcation and prediction. Oxford Statistical Science Series. Oxford University Press, 1st edition.

Zhou, X., Obuchowski, N., and McClish, D. (2011). Statistical Methods in Diagnostic Medicine. Wiley Series in Probability and Statistics. John Wiley & Sons, Hoboken, New Jersey, 2nd edition.

See also

Examples

data(Paired1) # Hypothetical study data 
b1 <- tab.paired(d=d, y1=y1, y2=y2, data=Paired1)
b2 <- acc.paired(b1)
print(b2)
#> Diagnostic accuracy of test 'y1'
#> 
#> (Estimates, standard errors and 95%-confidence intervals)
#> 
#>                  Est.         SE  Lower CL  Upper CL
#> Sensitivity 0.8802661 0.01528718 0.8503038 0.9102284
#> Specificity 0.6781609 0.02891782 0.6214830 0.7348388
#> PPV         0.8253638 0.01731081 0.7914353 0.8592924
#> NPV         0.7662338 0.02784617 0.7116563 0.8208113
#> 
#>            Est.  SE (log)  Lower CL  Upper CL
#> PDLR  2.7351124 0.0915147 2.2860079 3.2724472
#> NDLR  0.1765568 0.1346088 0.1356142 0.2298601
#> 
#> ----------------------------------------------------------
#> Diagnostic accuracy of test 'y2'
#> 
#> (Estimates, standard errors and 95%-confidence intervals)
#> 
#>                  Est.         SE  Lower CL  Upper CL
#> Sensitivity 0.7560976 0.02022128 0.7164646 0.7957305
#> Specificity 0.7969349 0.02490054 0.7481307 0.8457390
#> PPV         0.8654822 0.01718980 0.8317908 0.8991736
#> NPV         0.6540881 0.02667395 0.6018081 0.7063680
#> 
#>            Est.  SE (log)  Lower CL  Upper CL
#> PDLR  3.7234238 0.1255060 2.9114648 4.7618247
#> NDLR  0.3060507 0.0885996 0.2572629 0.3640906