Differences in Diagnostic Likelihood Ratios

Performs a test for differences in (positive and negative) diagnostic likelihood ratios (DLRs) of two binary diagnostic tests in a paired study using a regression model approach proposed by Gu and Pepe (2009).

Usage

dlr.regtest(tab, alpha)

Arguments

tab

An object of class tab.paired.

alpha

Significance level alpha for 100(1-alpha)%-confidence intervals, the default is 0.05.

Details

The null hypothesis rDLR = DLR of Test 1 / DLR of Test 2 = 1 is tested with respect to both positive and negative DLRs of the two diagnostic tests.

This function calls DLR, a general implementation of the method proposed by Gu and Pepe (2009).

Value

A list containing

pdlr

A list with test1 (the positive DLR of test 1), test2 (the positive DLR of test 2), ratio (the ratio of positive DLRs, computed as test1/test2, se.log (the standard error of the logarithm of ratio), the test.statistic and the corresponding p.value.

ndlr

A list with test1 (the negative DLR of test 1), test2 (the negative DLR of test 2), ratio (the ratio of negative DLRs, computed as test1/test2, se.log (the standard error of the logarithm of ratio), the test.statistic and the corresponding p.value.

alpha

The significance level alpha used to compute 100(1-alpha)%-confidence intervals for the ratio of positive and negative DLRs, the default is 0.05.

method

The name of the method used to compare the positive and negative DLRs, here “diagnostic likelihood regression model (regtest)”.

References

Gu, W. and Pepe, M. S. (2009). Estimating the capacity for improvement in risk prediction with a marker. Biostatistics, 10(1):172-86.

See also

Examples

data(Paired1) # Hypothetical study data
ptab <- tab.paired(d=d, y1=y1, y2=y2, data=Paired1)
ptab
#> Two binary diagnostic tests (paired design)
#> 
#> Test1: 'y1'
#> Test2: 'y2'
#> 
#> Diseased:
#>            Test1 pos. Test1 neg. Total
#> Test2 pos.        319         22   341
#> Test2 neg.         78         32   110
#> Total             397         54   451
#> 
#> Non-diseased:
#>            Test1 pos. Test1 neg. Total
#> Test2 pos.         31         22    53
#> Test2 neg.         53        155   208
#> Total              84        177   261
#> 
dlr.results <- dlr.regtest(ptab)
str(dlr.results)
#> List of 4
#>  $ pdlr  :List of 8
#>   ..$ test1         : num 2.74
#>   ..$ test2         : num 3.72
#>   ..$ ratio         : num 0.735
#>   ..$ se.log        : num 0.133
#>   ..$ test.statistic: num -2.33
#>   ..$ p.value       : num 0.02
#>   ..$ lcl           : num 0.566
#>   ..$ ucl           : num 0.953
#>  $ ndlr  :List of 8
#>   ..$ test1         : num 0.177
#>   ..$ test2         : num 0.306
#>   ..$ ratio         : num 0.577
#>   ..$ se.log        : num 0.137
#>   ..$ test.statistic: num -4
#>   ..$ p.value       : num 6.22e-05
#>   ..$ lcl           : num 0.441
#>   ..$ ucl           : num 0.755
#>  $ alpha : num 0.05
#>  $ method: chr "DLR regression model (regtest)"
dlr.results
#> $pdlr
#> $pdlr$test1
#> [1] 2.735112
#> 
#> $pdlr$test2
#> [1] 3.723424
#> 
#> $pdlr$ratio
#> [1] 0.7345692
#> 
#> $pdlr$se.log
#> [1] 0.1326086
#> 
#> $pdlr$test.statistic
#> [1] -2.326177
#> 
#> $pdlr$p.value
#> [1] 0.0200091
#> 
#> $pdlr$lcl
#> [1] 0.5664428
#> 
#> $pdlr$ucl
#> [1] 0.9525973
#> 
#> 
#> $ndlr
#> $ndlr$test1
#> [1] 0.1765568
#> 
#> $ndlr$test2
#> [1] 0.3060507
#> 
#> $ndlr$ratio
#> [1] 0.5768875
#> 
#> $ndlr$se.log
#> [1] 0.137376
#> 
#> $ndlr$test.statistic
#> [1] -4.004396
#> 
#> $ndlr$p.value
#> [1] 6.217627e-05
#> 
#> $ndlr$lcl
#> [1] 0.4407136
#> 
#> $ndlr$ucl
#> [1] 0.7551371
#> 
#> 
#> $alpha
#> [1] 0.05
#> 
#> $method
#> [1] "DLR regression model (regtest)"
#>