Print “tab.paired”-Object

Prints objects of class tab.paired in an easy-to-read form (S3method).

Usage

# S3 method for tab.paired
print(x, ...)

Arguments

x

An object of class tab.paired.

...

Additional arguments (usually not required).

Value

Creates a list object from parts of its input that is then printed in a tabular layout.

See also

Examples

data(Paired1) # Hypothetical study data 
b <- tab.paired(d=d, y1=y1, y2=y2, data=Paired1)
print(b)
#> Two binary diagnostic tests (paired design)
#> 
#> Test1: 'y1'
#> Test2: 'y2'
#> 
#> Diseased:
#>            Test1 pos. Test1 neg. Total
#> Test2 pos.        319         22   341
#> Test2 neg.         78         32   110
#> Total             397         54   451
#> 
#> Non-diseased:
#>            Test1 pos. Test1 neg. Total
#> Test2 pos.         31         22    53
#> Test2 neg.         53        155   208
#> Total              84        177   261
#>