Compute predictive values for theoretical prevalences

It is often of interest to estimate predictive values assuming the test were applied to a population with a different prevalence of the disease. Projected predictive values may be calculated using Bayes theorem and the relation between predictive values and diagnostic likelihood ratios can be used to derived corresponding confidence intervals.

Usage

pv.prev(pi, acc)

Arguments

pi

A theoretical prevalence of the disease (proportion).

acc

An object of class `acc.1test`.

Value

A vector containing the projected values.

Details

Predictive values, assuming a certain prevalence of the disease, are derived using the relation between predictive values and diagnostic likelihood ratios:

- PPV = 1 / (1 + (1 / pi - 1) / pDLR) - NPV = 1 / (1 + (1 / (1 / pi - 1)) / nDLR).

See Newcombe RG (2013). Confidence Intervals for Proportions and Related Measures of Effect Size. Chapman and Hall/ CRC Biostatistics Series (chapters 12.3+5 and 14.9).

The alpha-level of (1-alpha)-confidence intervals is inherited from `acc.1test`.

See also

Examples

data(Paired1) # Hypothetical study data 
a1 <- tab.1test(d=d, y=y1, data=Paired1)
a2 <- acc.1test(a1, alpha = 0.05)
pv.prev(pi=0.2, acc=a2)
#>        pi   ppv_est   ppv_lcl   ppv_ucl   npv_est   npv_lcl   npv_ucl 
#> 0.2000000 0.4060975 0.3636661 0.4499788 0.4139116 0.3516837 0.4790149 
pv.prev(pi=0.5, acc=a2)
#>        pi   ppv_est   ppv_lcl   ppv_ucl   npv_est   npv_lcl   npv_ucl 
#> 0.5000000 0.7322704 0.6956794 0.7659421 0.1500623 0.1194193 0.1868994