Read in “tab.1test”-Objects

Reads in objects of class tab.1test using cell frequencies.

Usage

read.tab.1test(a, b, c, d, testname, ...)

Arguments

a

The number of diseased subjects with a positive test.

b

The number of non-diseased subjects with a positive test.

c

The number of diseased subjects with a negative test.

d

The number of non-diseased subjects with a negative test.

testname

An optional vector specifying the name of the diagnostic test, e.g. c("Test A"). If not supplied, the corresponding variable name is used as testname.

...

Additional arguments (usually not required).

Value

Returns a list of class tab.1test containing:

tab.1test

A contingency table (matrix) of test results.

Diseased	Non-diseased	Total

Test pos.	a	b	a+b

Test neg.	c	d	c+d

Total	a+c	b+d	a+b+c+d

testname

The name of the diagnostic test.

Note

Objects of class tab.1test are required as arguments for acc.1test, a function to compute the accuracy of a binary diagnostic test.

See also

Examples

read.t1 <- read.tab.1test(321, 51, 730, 272, testname="Test1")
class(read.t1)
#> [1] "tab.1test"
read.t1
#> Binary diagnostic test 'Test1'
#> 
#>           Diseased Non-diseased Total
#> Test pos.      321           51   372
#> Test neg.      730          272  1002
#> Total         1051          323  1374
acc.1test(read.t1)
#> Diagnostic accuracy of test 'Test1'
#> 
#> (Estimates, standard errors and Wald's asymptotic normal-based 95%-confidence intervals)
#> 
#>                  Est.         SE  Lower CL  Upper CL
#> Sensitivity 0.3054234 0.01420725 0.2775777 0.3332691
#> Specificity 0.8421053 0.02028924 0.8023391 0.8818714
#> PPV         0.8629032 0.01783295 0.8279513 0.8978552
#> NPV         0.2714571 0.01404896 0.2439216 0.2989925
#> 
#>            Est.   SE (log)  Lower CL  Upper CL
#> PDLR  1.9343482 0.13665891 1.4798251 2.5284766
#> NDLR  0.8248097 0.03160512 0.7752672 0.8775181