Tabulate Single Binary Diagnostic Test vs. Gold-Standard

Produces a contingency table of results from a single binary diagnostic test vs. the gold-standard results.

Usage

tab.1test(d, y, data = NULL, testname, ...)

Arguments

d

A numeric vector specifying the gold-standard results (1 = presence of disease, 0 = absence of disease).

y

A numeric vector specifying the results of the diagnostic test (1 = positive, 0 = negative).

data

An optional data frame, list or environment containing the required variables d and y.

testname

An optional character variable specifying the name of the diagnostic test, e.g. c("Test A"). If not supplied, the variable name is used as the testname.

...

Additional arguments (usually not required).

Value

Returns a list of class tab.1test:

tab.1test

A contingency table (matrix) of test results.

Diseased	Non-diseased	Total

Test pos.	...	...	...

Test neg.	...	...	...

Total	...	...	...

testname

The name of the diagnostic test.

Note

Objects of class tab.1test are required as arguments for acc.1test, a function to compute the accuracy of a binary diagnostic test.

See also

Examples

data(Paired1) # Hypothetical study data 
a <- tab.1test(d=d, y=y1, data=Paired1)
str(a)
#> List of 2
#>  $ tab.1test: num [1:3, 1:3] 397 54 451 84 177 261 481 231 712
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:3] "Test pos." "Test neg." "Total"
#>   .. ..$ : chr [1:3] "Diseased" "Non-diseased" "Total"
#>  $ testname : chr "y1"
#>  - attr(*, "class")= chr "tab.1test"
a$tab.1test
#>           Diseased Non-diseased Total
#> Test pos.      397           84   481
#> Test neg.       54          177   231
#> Total          451          261   712
a
#> Binary diagnostic test 'y1'
#> 
#>           Diseased Non-diseased Total
#> Test pos.      397           84   481
#> Test neg.       54          177   231
#> Total          451          261   712